ml-service/requirements.txt
flask==3.1.0
./temp_chronos
pandas
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
python-dateutil==2.8.2
typing_extensions

